+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ## Added
 
+- Add custom template tag to convert ISO 8601 duration to a specified time unit.
 - Add custom serializer field for duration field to format ISO 8601
 - Add effort duration field for the Course Model
 - Add phone number field for the User Model